make processing script into a package, update README.md
